Refactor to create a separate module for version_tracker
