+++ pyproject.toml
@@ -18,7 +18,7 @@ fasttrack = 'apps.fasttrack.cli:cli'
 run_python_step = 'etl.run_python_step:main'
 compare = 'etl.compare:cli'
 generate_graph = 'etl.to_graphviz:to_graphviz'
-version_tracker = 'etl.helpers:run_version_tracker_checks'
+version_tracker = 'etl.version_tracker:run_version_tracker_checks'
 etl-harmonize = 'etl.harmonize:harmonize'
 etl-match-variables = 'etl.match_variables:main_cli'
 etl-translate-varmap = 'etl.variable_mapping_translate:main_cli'
